documentation typo, move everything in compositor namespace
