+++ com.unity.render-pipelines.high-definition/Editor/Compositor/SerializedCompositionLayer.cs
@@ -1,9 +1,9 @@
 using System.Collections.Generic;
 
 using UnityEditor;
-using UnityEngine.Rendering.HighDefinition;
+using UnityEngine.Rendering.HighDefinition.Compositor;
 
-namespace UnityEditor.Rendering.HighDefinition
+namespace UnityEditor.Rendering.HighDefinition.Compositor
 {
     internal class SerializedCompositionLayer
     {
