+++ com.unity.render-pipelines.high-definition/Editor/Compositor/SeriallizedShaderProperty.cs
@@ -1,6 +1,6 @@
 using UnityEditor;
 
-namespace UnityEditor.Rendering.HighDefinition
+namespace UnityEditor.Rendering.HighDefinition.Compositor
 {
     internal class SerializedShaderProperty
     {
