+++ com.unity.render-pipelines.high-definition/Runtime/Compositor/CustomClear.cs
@@ -3,7 +3,7 @@ using UnityEngine.Rendering.HighDefinition;
 using UnityEngine.Rendering;
 using UnityEngine.Experimental.Rendering;
 
-namespace UnityEngine.Rendering.HighDefinition
+namespace UnityEngine.Rendering.HighDefinition.Compositor
 {
     internal class CustomClear : CustomPass
     {
